Handle error in streaming call.
DeepInfra does not return an error code when streaming the response. The
response has a HTTP 200 code, but the event body has JSON that doesn't
match the values for the event type. This JSON is not quite the same as
the JSON from it returns in the non-streaming case. It has a different
top-leve structure, and true error is in a *almost* json encoded
payload. The payload has a prefix of "ConnectionError: ".